Neovim *almost* working; but LSP broken
So the cargo cult of the vcsh_nvim submodule into Home Manager
almost worked, but sadly the LSP (language server protocol) is
very much broken. This is a shame as it was so close. However,
the break is a bit fundamental, as the LSP servers haskell
(hls) and rust just can't work as the :LspInstall plugin downloads
binaries that link to unexpected places.  So basically broken.

This commit just marks where it got to, prior to moving the neovim
config to be more of a mix of nix and config files.