Add Neovim plugins, tree-sitter and LSP configs
This is a mostly complete Neovim setup.  I think that spelling is
still missing (in terms of libraries), but LSP, tree-sitter and most
things are now setup using the vcsh_nvim submodule.  It required a
little bit of work in the submodule repo to make it work with both Nix
and non-Nix systems.