Switch home-manager to a module in main nixos config
I had to switch the home-manager standalone installation to
a module in the nixos system configuration.nix as I used
`nix profile` and that breaks standalone home-manager installations
on nixos [1].  A pain, but difficult to resolve otherwise.  I've kept
the apply-user and update-user scripts for use on systems where
home-manager is standalone (which is currently none). They may
get removed.

I also ran into an issue [2] on home-manager that required an option
to be forced.

[1] https://discourse.nixos.org/t/depreciate-the-use-of-nix-env-to-install-packages/20139/32
[2] https://github.com/nix-community/home-manager/issues/1118